+++ CoreAlgorithms/test/unit/src/au/gov/asd/tac/constellation/plugins/algorithms/sna/global/ComponentsCountNGTest.java
@@ -15,11 +15,10 @@
  */
 package au.gov.asd.tac.constellation.plugins.algorithms.sna.global;
 
-import au.gov.asd.tac.constellation.plugins.algorithms.sna.global.ComponentCountPlugin;
 import au.gov.asd.tac.constellation.plugins.algorithms.sna.SnaConcept;
 import au.gov.asd.tac.constellation.graph.StoreGraph;
-import au.gov.asd.tac.constellation.schema.Schema;
-import au.gov.asd.tac.constellation.schema.SchemaFactoryUtilities;
+import au.gov.asd.tac.constellation.graph.schema.Schema;
+import au.gov.asd.tac.constellation.graph.schema.SchemaFactoryUtilities;
 import au.gov.asd.tac.constellation.plugins.PluginExecution;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
 import au.gov.asd.tac.constellation.graph.schema.analytic.AnalyticSchemaFactory;