+++ CoreGraphFile/test/unit/src/au/gov/asd/tac/constellation/graph/file/io/IONGTest.java
@@ -42,6 +42,22 @@ import org.testng.annotations.Test;
  */
 public class IONGTest {
 
+    @BeforeClass
+    public static void setUpClass() throws Exception {
+    }
+
+    @AfterClass
+    public static void tearDownClass() throws Exception {
+    }
+
+    @BeforeMethod
+    public void setUpMethod() throws Exception {
+    }
+
+    @AfterMethod
+    public void tearDownMethod() throws Exception {
+    }
+
     /**
      * Saving and loading an attribute with a non-null default value.
      */
@@ -134,21 +150,5 @@ public class IONGTest {
         } finally {
             rg.release();
         }
-    }
-
-    @BeforeClass
-    public static void setUpClass() throws Exception {
-    }
-
-    @AfterClass
-    public static void tearDownClass() throws Exception {
-    }
-
-    @BeforeMethod
-    public void setUpMethod() throws Exception {
-    }
-
-    @AfterMethod
-    public void tearDownMethod() throws Exception {
     }
 }