+++ CoreOpenGLDisplay/test/unit/src/au/gov/asd/tac/constellation/visual/opengl/DummyVisualAccess.java
@@ -17,11 +17,12 @@ package au.gov.asd.tac.constellation.visual.opengl;
 
 import au.gov.asd.tac.constellation.utilities.camera.Camera;
 import au.gov.asd.tac.constellation.utilities.color.ConstellationColor;
-import au.gov.asd.tac.constellation.graph.utilities.display.VisualAccess;
-import au.gov.asd.tac.constellation.graph.utilities.display.VisualChange;
 import au.gov.asd.tac.constellation.utilities.visual.DrawFlags;
 import au.gov.asd.tac.constellation.utilities.graphics.Vector3f;
 import au.gov.asd.tac.constellation.utilities.visual.LineStyle;
+import au.gov.asd.tac.constellation.utilities.visual.VisualAccess;
+import au.gov.asd.tac.constellation.utilities.visual.VisualAccess.ConnectionDirection;
+import au.gov.asd.tac.constellation.utilities.visual.VisualChange;
 import java.util.Collections;
 import java.util.List;