+++ CoreTesting/test/unit/src/au/gov/asd/tac/constellation/testing/integration/SchemaElementTypeNGTest.java
@@ -15,9 +15,9 @@
  */
 package au.gov.asd.tac.constellation.testing.integration;
 
-import au.gov.asd.tac.constellation.schema.SchemaElementType;
-import au.gov.asd.tac.constellation.schema.SchemaVertexType;
 import au.gov.asd.tac.constellation.graph.schema.analytic.concept.AnalyticConcept;
+import au.gov.asd.tac.constellation.graph.schema.type.SchemaElementType;
+import au.gov.asd.tac.constellation.graph.schema.type.SchemaVertexType;
 import au.gov.asd.tac.constellation.utilities.color.ConstellationColor;
 import static org.testng.Assert.assertEquals;
 import org.testng.annotations.AfterClass;